fix(connect,iOS) : add ifdef for when SignInWithApple not installed
