+++ Assets/Plugins/Source/Editor/com.playeveryware.eos-Editor.asmdef
@@ -4,8 +4,7 @@
     "references": [
         "com.playeveryware.eos",
         "Unity.EditorCoroutines.Editor",
-        "com.playeveryware.eos.essential",
-        "com.playeveryware.eos.samples.appleauth.editor"
+        "com.playeveryware.eos.essential"
     ],
     "includePlatforms": [
         "Editor"
