+++ Assets/Plugins/iOS/Editor/SignInWithApple/com.playeveryware.eos.samples.appleauth.editor.asmdef.meta
