+++ Assets/Scripts/SignInWithApple/com.playeveryware.eos.samples.appleauth.asmdef
